requirements.txt
moviepy
opencv-python 
pillow 
ffmpeg-python